fix($grab-one): frame rate param is mandatory  #1
This is required already for skipping past black frames, and if a timecode is specified instead of an integer for seek time.  While both of these are (or will be) optional paths, it seems like other feature will make it more likely that this value will be required.  To keep things simple, let's always require a frame rate.  We can revisit this decision later if necessary.